Support the use of GoBilda pinpoint odometry for localization.
Create calibrator for figuring out the position of the odometry pods. This will allow for manual rotations of robot, rotations using drive-train, and using april tags for fine tuning in case there is lateral displacement.

Introduce a MecanumWiring within Drives to save the wiring configuration.  This simplifies creation of the mecanum drive train.

Did complete revamp of nomenclature of Config classes across the framework and updated framework principles. This iteration standardized on having a defaults() method, and use of withX() nomenclature.  (More cleanup with config coming up.)